refactor: rename components and directories to better reflect app labels
