+++ client/src/components/Navbar/Navbar.tsx
@@ -67,7 +67,7 @@ const NavbarComponent: React.FC = () => {
             <Link to="/dash">Function Performance</Link>
           </li>
           <li>
-            <Link to="/cloudwatchmetrics">Cloudwatch Metrics</Link>
+            <Link to="/cloudwatchmetrics">CloudWatch Metrics</Link>
           </li>
           <li>
             <Link to="/">Configuration</Link>
