+++ client/src/components/Row/Row.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import RowComponent from "./RowComponent";
+import RowComponent from "./Row";
 
 describe("RowComponent", () => {
   // Define sample props that represent typical, non-edge-case values.
