+++ client/src/pages/CloudWatchMetrics/Throttle/Throttle.tsx
