+++ client/src/pages/CloudWatchMetrics/TotalDuration/TotalDuration.tsx
