+++ client/src/pages/FunctionPerformance/AverageBilledDuration/AverageBilledDuration.test.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import AvgBilledDurGraph from "./AvgBilledDurGraphComponent";
+import AvgBilledDurGraph from "./AverageBilledDuration";
 import type { Bar as ImportedBar } from "react-chartjs-2";
 
 type BarProps = React.ComponentProps<typeof ImportedBar>;
